Import and Convert Time-Series Data into Wide Format Sequence Data

Imports time-series data as sequence data by discretizing the input data. Various methods for discretization are available including gaussian mixtures, K-means clustering and kernel density based binning.

Usage

import_ts(
  data,
  id_col,
  value_col,
  n_states,
  method = "kmeans",
  unused_fn = dplyr::first,
  ...
)

Arguments

data

A data.frame containing time-series data in long format.

id_col

An optional character string naming the column that contains the unique IDs.

value_col

A character string naming the column that contains the data values.

n_states

An integer specifying the number of states.

method

A character string defining the discretization method to use. The available options are:

width: for equal width binning.

quantile: for quantile-based binning.

kde: for binning based on kernel density estimation.

gaussian: for a Gaussian mixture model.

kmeans: for K-means clustering (the default).

unused_fn

How to handle extra columns when pivoting to wide format. See tidyr::pivot_wider(). The default is to keep all columns and to use the first value.

...

Additional arguments passed to the discretization method (stats::kmeans() for kmeans, stats::density() and pracma::findpeaks() for kde, and mclust::Mclust() for gaussian).

Value

A data.frame in wide format containing the sequence data.

See also