Print Bootstrap Results

Print Bootstrap Results

Usage

# S3 method for class 'tna_bootstrap'
print(x, digits = getOption("digits"), type = "both", ...)

Arguments

x

A tna_bootstrap object.

digits

An integer giving the minimal number of significant digits to print.

type

A character vector giving the type of edges to print. The default option "both" prints both statistically significant and non-significant edges, "sig" prints only significant edges, and "nonsig" prints only the non-significant edges.

...

Ignored.

Value

x (invisibly).

See also

Examples

model <- tna(group_regulation)
# Small number of iterations for CRAN
boot <- bootstrap(model, iter = 10)
print(boot)
#> Non-significant Edges
#> 
#>          from         to       weight    p_value     ci_lower    ci_upper
#> 2       adapt   cohesion 0.0029498525 0.45454545 9.732264e-04 0.004144691
#> 3       adapt  consensus 0.0047400853 0.27272727 3.267040e-03 0.006047037
#> 4       adapt coregulate 0.0162436548 0.18181818 1.160189e-02 0.019619738
#> 5       adapt    discuss 0.0713743356 0.09090909 6.392616e-02 0.073950367
#> 6       adapt    emotion 0.0024673951 0.36363636 5.428535e-04 0.002784198
#> 7       adapt    monitor 0.0111653873 0.36363636 6.743735e-03 0.012781638
#> 8       adapt       plan 0.0009745006 0.54545455 6.381248e-06 0.001143400
#> 9       adapt  synthesis 0.2346625767 0.09090909 2.071694e-01 0.261940431
#> 10   cohesion      adapt 0.2730844794 0.09090909 2.162267e-01 0.301934840
#> 11   cohesion   cohesion 0.0271386431 0.09090909 2.177139e-02 0.030406966
#> 12   cohesion  consensus 0.0148522673 0.09090909 1.202293e-02 0.016466791
#> 13   cohesion coregulate 0.0360406091 0.09090909 2.807131e-02 0.042092344
#> 14   cohesion    discuss 0.0475828904 0.09090909 4.236954e-02 0.049973429
#> 15   cohesion    emotion 0.3253436729 0.09090909 3.176190e-01 0.338732383
#> 16   cohesion    monitor 0.0558269365 0.09090909 4.535463e-02 0.062932181
#> 17   cohesion       plan 0.0251745980 0.09090909 2.300836e-02 0.028487353
#> 18   cohesion  synthesis 0.0337423313 0.36363636 2.493834e-02 0.048774053
#> 19  consensus      adapt 0.4774066798 0.09090909 4.373891e-01 0.512139910
#> 20  consensus   cohesion 0.4979351032 0.09090909 4.799636e-01 0.521512904
#> 21  consensus  consensus 0.0820034761 0.09090909 8.070203e-02 0.087435148
#> 22  consensus coregulate 0.1345177665 0.09090909 1.231203e-01 0.146464193
#> 23  consensus    discuss 0.3211845103 0.09090909 3.144528e-01 0.337307616
#> 24  consensus    emotion 0.3204088826 0.09090909 3.132441e-01 0.328249775
#> 25  consensus    monitor 0.1591067690 0.09090909 1.496527e-01 0.184299243
#> 26  consensus       plan 0.2904011694 0.09090909 2.800919e-01 0.298987501
#> 27  consensus  synthesis 0.4662576687 0.09090909 4.136602e-01 0.508741433
#> 28 coregulate      adapt 0.0216110020 0.36363636 5.498760e-03 0.026962312
#> 29 coregulate   cohesion 0.1191740413 0.09090909 1.031090e-01 0.135547745
#> 30 coregulate  consensus 0.1877073787 0.09090909 1.846069e-01 0.192233671
#> 31 coregulate coregulate 0.0233502538 0.18181818 1.754917e-02 0.025054916
#> 32 coregulate    discuss 0.0842824601 0.09090909 7.394962e-02 0.092022060
#> 33 coregulate    emotion 0.0341910469 0.09090909 2.951808e-02 0.040518084
#> 34 coregulate    monitor 0.0579204466 0.09090909 5.148128e-02 0.063917078
#> 35 coregulate       plan 0.0172161767 0.09090909 1.552018e-02 0.019547927
#> 36 coregulate  synthesis 0.0444785276 0.36363636 2.404235e-02 0.058452381
#> 37    discuss      adapt 0.0589390963 0.36363636 3.506119e-02 0.078357861
#> 38    discuss   cohesion 0.0595870206 0.09090909 5.465161e-02 0.068834797
#> 39    discuss  consensus 0.1880233844 0.09090909 1.815152e-01 0.189317704
#> 40    discuss coregulate 0.2736040609 0.09090909 2.681602e-01 0.288165145
#> 41    discuss    discuss 0.1948873703 0.09090909 1.856051e-01 0.199266184
#> 42    discuss    emotion 0.1018681706 0.09090909 9.327653e-02 0.107333712
#> 43    discuss    monitor 0.3754361479 0.09090909 3.626083e-01 0.398594121
#> 44    discuss       plan 0.0678902063 0.09090909 6.467331e-02 0.072237260
#> 45    discuss  synthesis 0.0628834356 0.27272727 4.156304e-02 0.078984024
#> 46    emotion      adapt 0.1198428291 0.09090909 1.039156e-01 0.138463807
#> 47    emotion   cohesion 0.1156342183 0.09090909 1.004558e-01 0.127347360
#> 48    emotion  consensus 0.0726813083 0.09090909 6.823319e-02 0.077726360
#> 49    emotion coregulate 0.1720812183 0.09090909 1.595156e-01 0.191875681
#> 50    emotion    discuss 0.1057960010 0.09090909 9.635371e-02 0.110029799
#> 51    emotion    emotion 0.0768417342 0.09090909 7.138640e-02 0.088539327
#> 52    emotion    monitor 0.0907187718 0.09090909 7.796499e-02 0.099030012
#> 53    emotion       plan 0.1468247523 0.09090909 1.432132e-01 0.151475024
#> 54    emotion  synthesis 0.0705521472 0.09090909 6.454471e-02 0.077605779
#> 55    monitor      adapt 0.0333988212 0.27272727 1.330951e-02 0.038410278
#> 56    monitor   cohesion 0.0330383481 0.09090909 2.898782e-02 0.040203123
#> 57    monitor  consensus 0.0466108390 0.09090909 4.321725e-02 0.048980024
#> 58    monitor coregulate 0.0862944162 0.09090909 7.988124e-02 0.094789067
#> 59    monitor    discuss 0.0222728423 0.18181818 2.152112e-02 0.027979452
#> 60    monitor    emotion 0.0363059570 0.09090909 2.861202e-02 0.041474133
#> 61    monitor    monitor 0.0181437544 0.27272727 1.202435e-02 0.019377643
#> 62    monitor       plan 0.0755237941 0.09090909 7.020793e-02 0.082370606
#> 63    monitor  synthesis 0.0122699387 0.54545455 9.292382e-03 0.020665649
#> 64       plan      adapt 0.0157170923 0.45454545 3.687973e-03 0.022763354
#> 65       plan   cohesion 0.1410029499 0.09090909 1.354404e-01 0.150671060
#> 66       plan  consensus 0.3957971243 0.09090909 3.914580e-01 0.404962677
#> 67       plan coregulate 0.2390862944 0.09090909 2.159392e-01 0.253677461
#> 68       plan    discuss 0.0116426221 0.09090909 8.984599e-03 0.012592686
#> 69       plan    emotion 0.0997532605 0.09090909 9.462380e-02 0.110604688
#> 70       plan    monitor 0.2156315422 0.09090909 2.006563e-01 0.232680402
#> 71       plan       plan 0.3742082183 0.09090909 3.600033e-01 0.384534779
#> 72       plan  synthesis 0.0751533742 0.09090909 6.235753e-02 0.077466933
#> 74  synthesis   cohesion 0.0035398230 0.36363636 1.419670e-03 0.005057887
#> 75  synthesis  consensus 0.0075841365 0.27272727 4.925477e-03 0.009526902
#> 76  synthesis coregulate 0.0187817259 0.36363636 1.283109e-02 0.024736040
#> 77  synthesis    discuss 0.1409769679 0.09090909 1.356812e-01 0.148515551
#> 78  synthesis    emotion 0.0028198802 0.54545455 2.290562e-04 0.003881794
#> 79  synthesis    monitor 0.0160502442 0.63636364 7.762348e-03 0.021191263
#> 80  synthesis       plan 0.0017865844 0.63636364 1.157147e-03 0.002435449