Calculate Summary of Network Metrics for a grouped Transition Network

This function calculates a variety of network metrics for a tna object. It computes key metrics such as node and edge counts, network density, mean distance, strength measures, degree centrality, and reciprocity.

Usage

# S3 method for class 'group_tna'
summary(object, combined = TRUE, ...)

Arguments

object

A group_tna object.

combined

A logical indicating whether the summary results should be combined into a single data frame for all clusters (defaults to TRUE)

...

Ignored

Value

A summary.group_tna object which is a list of lists or a combined data.frame containing the following network metrics:

node_count: The total number of nodes.

edge_count: The total number of edges.

network_Density: The density of the network.

mean_distance: The mean shortest path length.

mean_out_strength: The mean out-strength of nodes.

sd_out_strength: The standard deviation of out-strength.

mean_in_strength: The mean in-strength of nodes.

sd_in_strength: The standard deviation of in-strength.

mean_out_degree: The mean out-degree of nodes.

sd_out_degree: The standard deviation of out-degree.

centralization_out_degree: The centralization of out-degree.

centralization_in_degree: The centralization of in-degree.

reciprocity: The reciprocity of the network.

Details

The function extracts the igraph network for each cluster and computes the following network metrics:

Node count: Total number of nodes in the network.

Edge count: Total number of edges in the network.

Network density: Proportion of possible edges that are present in the network.

Mean distance: The average shortest path length between nodes.

Mean and standard deviation of out-strength and in-strength: Measures of the total weight of outgoing and incoming edges for each node.

Mean and standard deviation of out-degree: The number of outgoing edges from each node.

Centralization of out-degree and in-degree: Measures of how centralized the network is based on the degrees of nodes.

Reciprocity: The proportion of edges that are reciprocated (i.e., mutual edges between nodes).

See also

Examples

group <- c(rep("High", 1000), rep("Low", 1000))
model <- group_model(group_regulation, group = group)
summary(model)
#> metric:
#> 
#>  [1] "Node Count"                  "Edge Count"                 
#>  [3] "Network Density"             "Mean Distance"              
#>  [5] "Mean Out-Strength"           "SD Out-Strength"            
#>  [7] "Mean In-Strength"            "SD In-Strength"             
#>  [9] "Mean Out-Degree"             "SD Out-Degree"              
#> [11] "Centralization (Out-Degree)" "Centralization (In-Degree)" 
#> [13] "Reciprocity"                
#> 
#> High:
#> 
#>  [1] 9.000000e+00 7.600000e+01 1.000000e+00 4.228677e-02 1.000000e+00
#>  [6] 9.141475e-01 1.000000e+00 7.850462e-17 8.444444e+00 1.130388e+00
#> [11] 4.687500e-02 4.687500e-02 9.565217e-01
#> 
#> Low:
#> 
#>  [1] 9.000000e+00 7.500000e+01 1.000000e+00 5.595597e-02 1.000000e+00
#>  [6] 7.185672e-01 1.000000e+00 3.925231e-17 8.333333e+00 8.660254e-01
#> [11] 6.250000e-02 6.250000e-02 9.411765e-01