Bootstrap Transition Networks from Sequence Data

Perform bootstrapping on transition networks created from sequence data stored in a tna object. Bootstrapped estimates of edge weights are returned with confidence intervals and significance testing.

Usage

bootstrap(x, ...)

# S3 method for class 'tna'
bootstrap(
  x,
  iter = 1000,
  level = 0.05,
  method = "stability",
  threshold,
  consistency_range = c(0.75, 1.25),
  ...
)

# S3 method for class 'group_tna'
bootstrap(
  x,
  iter = 1000,
  level = 0.05,
  method = "stability",
  threshold,
  consistency_range = c(0.75, 1.25),
  ...
)

Arguments

x

A tna or a group_tna object created from sequence data.

...

Ignored.

iter

An integer specifying the number of bootstrap samples to draw. Defaults to 1000.

level

A numeric value representing the significance level for hypothesis testing and confidence intervals. Defaults to 0.05.

method

A character string. This argument defines the bootstrap test statistic. The "stability" option (the default) compares edge weights against a range of "consistent" values defined by consistency_range. Weights that fall outside this range are considered insignificant. In other words, an edge is considered significant if its value is within the range in (1 - level) * 100% of the bootstrap samples. The "threshold" option instead compares the edge weights against a user-specified threshold value.

threshold

A numeric value to compare edge weights against. The default is the 10th percentile of the edge weights. Used only when method = "threshold".

consistency_range

A numeric vector of length 2. Determines how much the edge weights may deviate (multiplicatively) from their observed values (below and above) before they are considered insignificant. The default is c(0.75, 1.25) which corresponds to a symmetric 25% deviation range. Used only when method = "stability".

Value

A tna_bootstrap object which is a list containing the following elements:

weights_orig: The original edge weight matrix.

weights_sig: The matrix of significant transitions (those with estimated p-values below the significance level).

weights_mean: The mean weight matrix from the bootstrap samples.

weights_sd: The standard deviation matrix from the bootstrap samples.

ci_lower: The lower bound matrix of the confidence intervals for the edge weights.

ci_upper: The upper bound matrix of the confidence intervals for the edge weights.

p_values: The matrix of estimated p-values for the edge weights.

summary: A data.frame summarizing the edges, their weights, p-values, statistical significance and confidence intervals.

If x is a group_tna object, the output is a group_tna_bootstrap object, which is a list of tna_bootstrap objects.

Details

The function first computes the original edge weights for the specified cluster from the tna object. It then performs bootstrapping by resampling the sequence data and recalculating the edge weights for each bootstrap sample. The mean and standard deviation of the transitions are computed, and confidence intervals are derived. The function also estimates p-values for each edge and identifies significant edges based on the specified significance level. A matrix of significant edges (those with estimated p-values below the significance level) is generated. Additional statistics on removed edges (those not considered significant) are provided.

All results, including the original transition matrix, bootstrapped estimates, and summary statistics for removed edges, are returned in a structured list.

See also

Examples

model <- tna(group_regulation)
# Small number of iterations for CRAN
bootstrap(model, iter = 10)
#> Non-significant Edges
#> 
#>          from         to       weight    p_value     ci_lower    ci_upper
#> 2       adapt   cohesion 0.0029498525 0.72727273 0.0007146923 0.005853452
#> 3       adapt  consensus 0.0047400853 0.18181818 0.0031162999 0.005670046
#> 4       adapt coregulate 0.0162436548 0.27272727 0.0132977590 0.023854536
#> 5       adapt    discuss 0.0713743356 0.09090909 0.0638149019 0.077541151
#> 6       adapt    emotion 0.0024673951 0.54545455 0.0013932185 0.003958716
#> 7       adapt    monitor 0.0111653873 0.18181818 0.0084833480 0.016781683
#> 8       adapt       plan 0.0009745006 0.54545455 0.0006938210 0.001610183
#> 9       adapt  synthesis 0.2346625767 0.09090909 0.1997386662 0.265442566
#> 10   cohesion      adapt 0.2730844794 0.09090909 0.2438849341 0.302781017
#> 11   cohesion   cohesion 0.0271386431 0.18181818 0.0204571493 0.034753199
#> 12   cohesion  consensus 0.0148522673 0.09090909 0.0121010910 0.017028585
#> 13   cohesion coregulate 0.0360406091 0.18181818 0.0336605579 0.047170932
#> 14   cohesion    discuss 0.0475828904 0.09090909 0.0407473912 0.051709383
#> 15   cohesion    emotion 0.3253436729 0.09090909 0.3197919209 0.347638871
#> 16   cohesion    monitor 0.0558269365 0.18181818 0.0546041045 0.068622062
#> 17   cohesion       plan 0.0251745980 0.09090909 0.0214130553 0.027621960
#> 18   cohesion  synthesis 0.0337423313 0.18181818 0.0243012422 0.041056741
#> 19  consensus      adapt 0.4774066798 0.09090909 0.4362729741 0.511017683
#> 20  consensus   cohesion 0.4979351032 0.09090909 0.4815855147 0.511561367
#> 21  consensus  consensus 0.0820034761 0.09090909 0.0795857757 0.085315095
#> 22  consensus coregulate 0.1345177665 0.09090909 0.1278009206 0.142560246
#> 23  consensus    discuss 0.3211845103 0.09090909 0.3100289980 0.331382767
#> 24  consensus    emotion 0.3204088826 0.09090909 0.3056782519 0.334150476
#> 25  consensus    monitor 0.1591067690 0.09090909 0.1406519454 0.165914290
#> 26  consensus       plan 0.2904011694 0.09090909 0.2897859163 0.302470422
#> 27  consensus  synthesis 0.4662576687 0.09090909 0.4513647874 0.475278524
#> 28 coregulate      adapt 0.0216110020 0.54545455 0.0121493300 0.028801170
#> 29 coregulate   cohesion 0.1191740413 0.09090909 0.1081870658 0.135100592
#> 30 coregulate  consensus 0.1877073787 0.09090909 0.1843548395 0.195524913
#> 31 coregulate coregulate 0.0233502538 0.09090909 0.0195131497 0.028706760
#> 32 coregulate    discuss 0.0842824601 0.09090909 0.0770015788 0.089252192
#> 33 coregulate    emotion 0.0341910469 0.09090909 0.0307200308 0.037245795
#> 34 coregulate    monitor 0.0579204466 0.09090909 0.0481798000 0.064730096
#> 35 coregulate       plan 0.0172161767 0.09090909 0.0158511458 0.019165171
#> 36 coregulate  synthesis 0.0444785276 0.36363636 0.0313375538 0.055543135
#> 37    discuss      adapt 0.0589390963 0.27272727 0.0474711583 0.085000194
#> 38    discuss   cohesion 0.0595870206 0.18181818 0.0515186393 0.075129917
#> 39    discuss  consensus 0.1880233844 0.09090909 0.1822526683 0.194385937
#> 40    discuss coregulate 0.2736040609 0.09090909 0.2528697655 0.290699150
#> 41    discuss    discuss 0.1948873703 0.09090909 0.1820757887 0.206331967
#> 42    discuss    emotion 0.1018681706 0.09090909 0.0909250529 0.104716602
#> 43    discuss    monitor 0.3754361479 0.09090909 0.3515781781 0.388930476
#> 44    discuss       plan 0.0678902063 0.09090909 0.0609293156 0.070684171
#> 45    discuss  synthesis 0.0628834356 0.27272727 0.0394863054 0.080687712
#> 46    emotion      adapt 0.1198428291 0.09090909 0.0969844733 0.127203672
#> 47    emotion   cohesion 0.1156342183 0.09090909 0.1035633798 0.122529637
#> 48    emotion  consensus 0.0726813083 0.09090909 0.0667320936 0.078623799
#> 49    emotion coregulate 0.1720812183 0.09090909 0.1643910620 0.187999322
#> 50    emotion    discuss 0.1057960010 0.09090909 0.1016735093 0.110177555
#> 51    emotion    emotion 0.0768417342 0.09090909 0.0725146125 0.081013369
#> 52    emotion    monitor 0.0907187718 0.09090909 0.0794946880 0.101194839
#> 53    emotion       plan 0.1468247523 0.09090909 0.1415594269 0.156462951
#> 54    emotion  synthesis 0.0705521472 0.18181818 0.0632843893 0.093488294
#> 55    monitor      adapt 0.0333988212 0.27272727 0.0239102782 0.044347108
#> 56    monitor   cohesion 0.0330383481 0.09090909 0.0256797285 0.037207034
#> 57    monitor  consensus 0.0466108390 0.09090909 0.0416597861 0.047196375
#> 58    monitor coregulate 0.0862944162 0.09090909 0.0787334698 0.089710913
#> 59    monitor    discuss 0.0222728423 0.09090909 0.0196590606 0.025546163
#> 60    monitor    emotion 0.0363059570 0.09090909 0.0314658644 0.044426298
#> 61    monitor    monitor 0.0181437544 0.18181818 0.0146094279 0.025527662
#> 62    monitor       plan 0.0755237941 0.09090909 0.0713841143 0.081492917
#> 63    monitor  synthesis 0.0122699387 0.63636364 0.0064481930 0.023259576
#> 64       plan      adapt 0.0157170923 0.54545455 0.0138291810 0.027566625
#> 65       plan   cohesion 0.1410029499 0.09090909 0.1274385142 0.162001188
#> 66       plan  consensus 0.3957971243 0.09090909 0.3860720820 0.401296741
#> 67       plan coregulate 0.2390862944 0.09090909 0.2293979651 0.250658721
#> 68       plan    discuss 0.0116426221 0.18181818 0.0102359407 0.014716862
#> 69       plan    emotion 0.0997532605 0.09090909 0.0900343110 0.104770720
#> 70       plan    monitor 0.2156315422 0.09090909 0.2016383750 0.233177807
#> 71       plan       plan 0.3742082183 0.09090909 0.3607978815 0.372833260
#> 72       plan  synthesis 0.0751533742 0.18181818 0.0642758596 0.096421129
#> 74  synthesis   cohesion 0.0035398230 0.54545455 0.0017301136 0.004762985
#> 75  synthesis  consensus 0.0075841365 0.18181818 0.0063464408 0.009617559
#> 76  synthesis coregulate 0.0187817259 0.36363636 0.0112842922 0.024291955
#> 77  synthesis    discuss 0.1409769679 0.09090909 0.1372504389 0.150918653
#> 78  synthesis    emotion 0.0028198802 0.54545455 0.0017208090 0.003462984
#> 79  synthesis    monitor 0.0160502442 0.18181818 0.0126187487 0.019941206
#> 80  synthesis       plan 0.0017865844 0.36363636 0.0011731180 0.002271774

model <- group_tna(engagement_mmm)
# Small number of iterations for CRAN
bootstrap(model, iter = 10)
#> Cluster 1:
#> 
#> Non-significant Edges
#> 
#>         from         to     weight    p_value   ci_lower   ci_upper
#> 1     Active     Active 0.70615034 0.09090909 0.67289922 0.74825737
#> 2     Active    Average 0.51339286 0.18181818 0.46339108 0.63430012
#> 3     Active Disengaged 0.33333333 0.63636364 0.19031250 0.52279412
#> 4    Average     Active 0.29384966 0.09090909 0.25174263 0.32710078
#> 5    Average    Average 0.45982143 0.09090909 0.35433951 0.51307546
#> 6    Average Disengaged 0.38095238 0.36363636 0.24502262 0.49464286
#> 8 Disengaged    Average 0.02678571 0.54545455 0.00553225 0.04441871
#> 9 Disengaged Disengaged 0.28571429 0.54545455 0.15893665 0.38941176
#> 
#> Cluster 2:
#> 
#> Non-significant Edges
#> 
#>         from         to     weight    p_value   ci_lower   ci_upper
#> 1     Active     Active 0.49242424 0.09090909 0.45029788 0.54403767
#> 2     Active    Average 0.32495164 0.09090909 0.29270069 0.36508528
#> 3     Active Disengaged 0.33333333 0.09090909 0.26658958 0.38143797
#> 4    Average     Active 0.44191919 0.09090909 0.39486021 0.47819492
#> 5    Average    Average 0.58607350 0.09090909 0.55025978 0.62359705
#> 6    Average Disengaged 0.57333333 0.09090909 0.55004633 0.68263145
#> 7 Disengaged     Active 0.06565657 0.18181818 0.05208170 0.08072051
#> 8 Disengaged    Average 0.08897485 0.18181818 0.06874414 0.11147549
#> 9 Disengaged Disengaged 0.09333333 0.81818182 0.05077897 0.13723442
#> 
#> Cluster 3:
#> 
#> Non-significant Edges
#> 
#>         from         to     weight    p_value   ci_lower   ci_upper
#> 1     Active     Active 0.30067568 0.09090909 0.24733897 0.32707386
#> 2     Active    Average 0.15940767 0.09090909 0.13677306 0.16762306
#> 3     Active Disengaged 0.06578947 0.18181818 0.05189673 0.08148381
#> 4    Average     Active 0.60810811 0.09090909 0.56903674 0.62852029
#> 5    Average    Average 0.57317073 0.09090909 0.55702593 0.59861532
#> 6    Average Disengaged 0.47076023 0.09090909 0.44656634 0.50002019
#> 7 Disengaged     Active 0.09121622 0.36363636 0.08740210 0.13389860
#> 8 Disengaged    Average 0.26742160 0.09090909 0.24981447 0.28596611
#> 9 Disengaged Disengaged 0.46345029 0.09090909 0.43275505 0.48242334