Print a Bootstrap Summary for a Grouped Transition Network Model

Print a Bootstrap Summary for a Grouped Transition Network Model

Usage

# S3 method for class 'summary.group_tna_bootstrap'
print(x, ...)

Arguments

x

A summary.group_tna_bootstrap object.

...

Arguments passed to the generic print method.

Value

x (invisibly).

See also

Examples

model <- group_model(engagement_mmm)
# Low number of iteration for CRAN
boot <- bootstrap(model, iter = 10)
print(summary(boot))
#>        group       from         to     weight    p_value   sig   ci_lower
#> 1  Cluster 1     Active     Active 0.70615034 0.09090909 FALSE 0.66940990
#> 2  Cluster 1     Active    Average 0.51339286 0.09090909 FALSE 0.40649716
#> 3  Cluster 1     Active Disengaged 0.33333333 0.54545455 FALSE 0.14387255
#> 4  Cluster 1    Average     Active 0.29384966 0.09090909 FALSE 0.26605181
#> 5  Cluster 1    Average    Average 0.45982143 0.18181818 FALSE 0.39415115
#> 6  Cluster 1    Average Disengaged 0.38095238 0.45454545 FALSE 0.16829365
#> 7  Cluster 1 Disengaged    Average 0.02678571 0.81818182 FALSE 0.00272154
#> 8  Cluster 1 Disengaged Disengaged 0.28571429 0.54545455 FALSE 0.18514610
#> 9  Cluster 2     Active     Active 0.49242424 0.09090909 FALSE 0.43969352
#> 10 Cluster 2     Active    Average 0.32495164 0.09090909 FALSE 0.26679314
#> 11 Cluster 2     Active Disengaged 0.33333333 0.27272727 FALSE 0.23669379
#> 12 Cluster 2    Average     Active 0.44191919 0.09090909 FALSE 0.38226570
#> 13 Cluster 2    Average    Average 0.58607350 0.09090909 FALSE 0.55302246
#> 14 Cluster 2    Average Disengaged 0.57333333 0.09090909 FALSE 0.53134384
#> 15 Cluster 2 Disengaged     Active 0.06565657 0.09090909 FALSE 0.05355145
#> 16 Cluster 2 Disengaged    Average 0.08897485 0.09090909 FALSE 0.07334572
#> 17 Cluster 2 Disengaged Disengaged 0.09333333 0.45454545 FALSE 0.05132510
#> 18 Cluster 3     Active     Active 0.30067568 0.09090909 FALSE 0.25710107
#> 19 Cluster 3     Active    Average 0.15940767 0.09090909 FALSE 0.13904481
#> 20 Cluster 3     Active Disengaged 0.06578947 0.09090909 FALSE 0.05304883
#> 21 Cluster 3    Average     Active 0.60810811 0.09090909 FALSE 0.58669008
#> 22 Cluster 3    Average    Average 0.57317073 0.09090909 FALSE 0.54910501
#> 23 Cluster 3    Average Disengaged 0.47076023 0.09090909 FALSE 0.45474700
#> 24 Cluster 3 Disengaged     Active 0.09121622 0.45454545 FALSE 0.07011639
#> 25 Cluster 3 Disengaged    Average 0.26742160 0.09090909 FALSE 0.23304284
#> 26 Cluster 3 Disengaged Disengaged 0.46345029 0.09090909 FALSE 0.44556511
#>      ci_upper
#> 1  0.73394819
#> 2  0.57146537
#> 3  0.53555556
#> 4  0.33059010
#> 5  0.58709965
#> 6  0.46237268
#> 7  0.04310395
#> 8  0.40255602
#> 9  0.55628944
#> 10 0.35463412
#> 11 0.39722817
#> 12 0.48923643
#> 13 0.63903545
#> 14 0.65816757
#> 15 0.07259875
#> 16 0.09506313
#> 17 0.12630952
#> 18 0.32829479
#> 19 0.17800598
#> 20 0.07497418
#> 21 0.66945248
#> 22 0.60324723
#> 23 0.49260644
#> 24 0.12019478
#> 25 0.29156039
#> 26 0.48833278