Print a Bootstrap Summary for a Grouped Transition Network Model

Print a Bootstrap Summary for a Grouped Transition Network Model

Usage

# S3 method for class 'summary.group_tna_bootstrap'
print(x, ...)

Arguments

x

A summary.group_tna_bootstrap object.

...

Arguments passed to the generic print method.

Value

x (invisibly).

See also

Examples

model <- group_model(engagement_mmm)
# Low number of iteration for CRAN
boot <- bootstrap(model, iter = 10)
print(summary(boot))
#>        group       from         to     weight    p_value   sig   cr_lower
#> 1  Cluster 1     Active     Active 0.85985688 0.09090909 FALSE 0.64489266
#> 2  Cluster 1    Average     Active 0.31210322 0.09090909 FALSE 0.23407741
#> 3  Cluster 1 Disengaged     Active 0.04791061 0.09090909 FALSE 0.03593296
#> 4  Cluster 1     Active    Average 0.08919748 0.09090909 FALSE 0.06689811
#> 5  Cluster 1    Average    Average 0.54208478 0.09090909 FALSE 0.40656359
#> 6  Cluster 1 Disengaged    Average 0.16179397 0.09090909 FALSE 0.12134548
#> 7  Cluster 1     Active Disengaged 0.05094565 0.09090909 FALSE 0.03820923
#> 8  Cluster 1    Average Disengaged 0.14581200 0.09090909 FALSE 0.10935900
#> 9  Cluster 1 Disengaged Disengaged 0.79029542 0.09090909 FALSE 0.59272157
#> 10 Cluster 2     Active     Active 0.84090909 0.09090909 FALSE 0.63068182
#> 11 Cluster 2    Average     Active 0.09259259 0.36363636 FALSE 0.06944444
#> 12 Cluster 2 Disengaged     Active 0.15555556 0.27272727 FALSE 0.11666667
#> 13 Cluster 2     Active    Average 0.15909091 0.09090909 FALSE 0.11931818
#> 14 Cluster 2    Average    Average 0.62962963 0.09090909 FALSE 0.47222222
#> 15 Cluster 2 Disengaged    Average 0.51111111 0.09090909 FALSE 0.38333333
#> 16 Cluster 2    Average Disengaged 0.27777778 0.18181818 FALSE 0.20833333
#> 17 Cluster 2 Disengaged Disengaged 0.33333333 0.09090909 FALSE 0.25000000
#> 18 Cluster 3     Active     Active 0.58333333 0.09090909 FALSE 0.43750000
#> 19 Cluster 3    Average     Active 0.15277778 0.27272727 FALSE 0.11458333
#> 20 Cluster 3     Active    Average 0.12500000 0.72727273 FALSE 0.09375000
#> 21 Cluster 3    Average    Average 0.81944444 0.09090909 FALSE 0.61458333
#> 22 Cluster 3 Disengaged    Average 0.60000000 0.27272727 FALSE 0.45000000
#> 23 Cluster 3     Active Disengaged 0.29166667 0.18181818 FALSE 0.21875000
#> 24 Cluster 3    Average Disengaged 0.02777778 1.00000000 FALSE 0.02083333
#> 25 Cluster 3 Disengaged Disengaged 0.40000000 0.36363636 FALSE 0.30000000
#>      cr_upper    ci_lower   ci_upper
#> 1  1.07482109 0.853543322 0.86657782
#> 2  0.39012902 0.299904897 0.32329947
#> 3  0.05988826 0.044985503 0.05170664
#> 4  0.11149685 0.084160382 0.09355355
#> 5  0.67760598 0.529044361 0.55285185
#> 6  0.20224246 0.159106876 0.16841747
#> 7  0.06368206 0.049209095 0.05448622
#> 8  0.18226500 0.138229062 0.15412379
#> 9  0.98786928 0.780609795 0.79549810
#> 10 1.05113636 0.818287844 0.86076555
#> 11 0.11574074 0.086401563 0.15380414
#> 12 0.19444444 0.082414242 0.18753469
#> 13 0.19886364 0.139234450 0.18171216
#> 14 0.78703704 0.503808891 0.67010313
#> 15 0.63888889 0.500000000 0.61594659
#> 16 0.34722222 0.243495313 0.34238697
#> 17 0.41666667 0.255564648 0.34163158
#> 18 0.72916667 0.465301003 0.71294643
#> 19 0.19097222 0.121304259 0.21071747
#> 20 0.15625000 0.048363095 0.18816890
#> 21 1.02430556 0.781316845 0.83729901
#> 22 0.75000000 0.509375000 0.79843750
#> 23 0.36458333 0.222321429 0.35726087
#> 24 0.03472222 0.003082192 0.05269303
#> 25 0.50000000 0.201562500 0.49062500