Print a Bootstrap Summary for a Grouped Transition Network Model

Print a Bootstrap Summary for a Grouped Transition Network Model

Usage

# S3 method for class 'summary.group_tna_bootstrap'
print(x, ...)

Arguments

x

A summary.group_tna_bootstrap object.

...

Arguments passed to the generic print method.

Value

x (invisibly).

See also

Examples

model <- group_model(engagement_mmm)
# Low number of iteration for CRAN
boot <- bootstrap(model, iter = 10)
print(summary(boot))
#>        group       from         to     weight    p_value   sig   cr_lower
#> 1  Cluster 1     Active     Active 0.70615034 0.09090909 FALSE 0.52961276
#> 2  Cluster 1    Average     Active 0.51339286 0.09090909 FALSE 0.38504464
#> 3  Cluster 1 Disengaged     Active 0.33333333 0.54545455 FALSE 0.25000000
#> 4  Cluster 1     Active    Average 0.29384966 0.09090909 FALSE 0.22038724
#> 5  Cluster 1    Average    Average 0.45982143 0.09090909 FALSE 0.34486607
#> 6  Cluster 1 Disengaged    Average 0.38095238 0.36363636 FALSE 0.28571429
#> 7  Cluster 1    Average Disengaged 0.02678571 0.54545455 FALSE 0.02008929
#> 8  Cluster 1 Disengaged Disengaged 0.28571429 0.27272727 FALSE 0.21428571
#> 9  Cluster 2     Active     Active 0.49242424 0.09090909 FALSE 0.36931818
#> 10 Cluster 2    Average     Active 0.32495164 0.09090909 FALSE 0.24371373
#> 11 Cluster 2 Disengaged     Active 0.33333333 0.09090909 FALSE 0.25000000
#> 12 Cluster 2     Active    Average 0.44191919 0.09090909 FALSE 0.33143939
#> 13 Cluster 2    Average    Average 0.58607350 0.09090909 FALSE 0.43955513
#> 14 Cluster 2 Disengaged    Average 0.57333333 0.09090909 FALSE 0.43000000
#> 15 Cluster 2     Active Disengaged 0.06565657 0.18181818 FALSE 0.04924242
#> 16 Cluster 2    Average Disengaged 0.08897485 0.09090909 FALSE 0.06673114
#> 17 Cluster 2 Disengaged Disengaged 0.09333333 0.63636364 FALSE 0.07000000
#> 18 Cluster 3     Active     Active 0.30067568 0.09090909 FALSE 0.22550676
#> 19 Cluster 3    Average     Active 0.15940767 0.09090909 FALSE 0.11955575
#> 20 Cluster 3 Disengaged     Active 0.06578947 0.09090909 FALSE 0.04934211
#> 21 Cluster 3     Active    Average 0.60810811 0.09090909 FALSE 0.45608108
#> 22 Cluster 3    Average    Average 0.57317073 0.09090909 FALSE 0.42987805
#> 23 Cluster 3 Disengaged    Average 0.47076023 0.09090909 FALSE 0.35307018
#> 24 Cluster 3     Active Disengaged 0.09121622 0.18181818 FALSE 0.06841216
#> 25 Cluster 3    Average Disengaged 0.26742160 0.09090909 FALSE 0.20056620
#> 26 Cluster 3 Disengaged Disengaged 0.46345029 0.09090909 FALSE 0.34758772
#>      cr_upper   ci_lower   ci_upper
#> 1  0.88268793 0.68570721 0.73750868
#> 2  0.64174107 0.44542553 0.54914142
#> 3  0.41666667 0.15231481 0.42214286
#> 4  0.36731207 0.26249132 0.31429279
#> 5  0.57477679 0.43089851 0.52904255
#> 6  0.47619048 0.30750000 0.53522727
#> 7  0.03348214 0.01138251 0.04448052
#> 8  0.35714286 0.23852381 0.37395833
#> 9  0.61553030 0.45511967 0.54041872
#> 10 0.40618956 0.29622295 0.35656521
#> 11 0.41666667 0.26562121 0.40423295
#> 12 0.55239899 0.39956723 0.48327111
#> 13 0.73259188 0.55529738 0.62052180
#> 14 0.71666667 0.52224242 0.61898032
#> 15 0.08207071 0.04429803 0.07425608
#> 16 0.11121857 0.07187635 0.10824180
#> 17 0.11666667 0.05107877 0.15468182
#> 18 0.37584459 0.26713900 0.33236559
#> 19 0.19925958 0.14651283 0.16721112
#> 20 0.08223684 0.05582705 0.07876408
#> 21 0.76013514 0.55750000 0.64445488
#> 22 0.71646341 0.55028234 0.58466646
#> 23 0.58845029 0.46659471 0.49600365
#> 24 0.11402027 0.07722171 0.11544170
#> 25 0.33427700 0.25754343 0.28760613
#> 26 0.57931287 0.44376379 0.47179075