Print a Bootstrap Summary for a Grouped Transition Network Model

Print a Bootstrap Summary for a Grouped Transition Network Model

Usage

# S3 method for class 'summary.group_tna_bootstrap'
print(x, ...)

Arguments

x

A summary.group_tna_bootstrap object.

...

Arguments passed to the generic print method.

Value

x (invisibly).

See also

Examples

model <- group_model(engagement_mmm)
# Low number of iteration for CRAN
boot <- bootstrap(model, iter = 10)
print(summary(boot))
#>        group       from         to     weight p_value   sig   ci_lower
#> 1  Cluster 1     Active     Active 0.70615034     0.0  TRUE 0.67835249
#> 2  Cluster 1     Active    Average 0.51339286     0.0  TRUE 0.45532035
#> 3  Cluster 1     Active Disengaged 0.33333333     0.5 FALSE 0.13111111
#> 4  Cluster 1    Average     Active 0.29384966     0.0  TRUE 0.25710922
#> 5  Cluster 1    Average    Average 0.45982143     0.1 FALSE 0.33254321
#> 6  Cluster 1    Average Disengaged 0.38095238     0.4 FALSE 0.29953209
#> 7  Cluster 1 Disengaged    Average 0.02678571     0.8 FALSE 0.01046748
#> 8  Cluster 1 Disengaged Disengaged 0.28571429     0.5 FALSE 0.16887255
#> 9  Cluster 2     Active     Active 0.49242424     0.0  TRUE 0.42855905
#> 10 Cluster 2     Active    Average 0.32495164     0.0  TRUE 0.29526917
#> 11 Cluster 2     Active Disengaged 0.33333333     0.2 FALSE 0.26943849
#> 12 Cluster 2    Average     Active 0.44191919     0.0  TRUE 0.39460196
#> 13 Cluster 2    Average    Average 0.58607350     0.0  TRUE 0.53311155
#> 14 Cluster 2    Average Disengaged 0.57333333     0.0  TRUE 0.48849910
#> 15 Cluster 2 Disengaged     Active 0.06565657     0.0  TRUE 0.05871438
#> 16 Cluster 2 Disengaged    Average 0.08897485     0.0  TRUE 0.08288658
#> 17 Cluster 2 Disengaged Disengaged 0.09333333     0.4 FALSE 0.06035714
#> 18 Cluster 3     Active     Active 0.30067568     0.0  TRUE 0.27305656
#> 19 Cluster 3     Active    Average 0.15940767     0.0  TRUE 0.14080935
#> 20 Cluster 3     Active Disengaged 0.06578947     0.0  TRUE 0.05660476
#> 21 Cluster 3    Average     Active 0.60810811     0.0  TRUE 0.54676374
#> 22 Cluster 3    Average    Average 0.57317073     0.0  TRUE 0.54309424
#> 23 Cluster 3    Average Disengaged 0.47076023     0.0  TRUE 0.44891403
#> 24 Cluster 3 Disengaged     Active 0.09121622     0.4 FALSE 0.06223765
#> 25 Cluster 3 Disengaged    Average 0.26742160     0.0  TRUE 0.24328282
#> 26 Cluster 3 Disengaged Disengaged 0.46345029     0.0  TRUE 0.43856781
#>      ci_upper
#> 1  0.74289078
#> 2  0.62028856
#> 3  0.52279412
#> 4  0.32164751
#> 5  0.52549171
#> 6  0.59361111
#> 7  0.05084989
#> 8  0.38628247
#> 9  0.54515497
#> 10 0.38311015
#> 11 0.42997288
#> 12 0.50157268
#> 13 0.61912454
#> 14 0.61532283
#> 15 0.07776168
#> 16 0.10460399
#> 17 0.13534157
#> 18 0.34425028
#> 19 0.17977052
#> 20 0.07853011
#> 21 0.62952613
#> 22 0.59723645
#> 23 0.48677347
#> 24 0.11231604
#> 25 0.30180036
#> 26 0.48133548