Print Bootstrap Results

Print Bootstrap Results

Usage

# S3 method for class 'tna_bootstrap'
print(x, digits = getOption("digits"), type = "both", ...)

Arguments

x

A tna_bootstrap object.

digits

An integer giving the minimal number of significant digits to print.

type

A character vector giving the type of edges to print. The default option "both" prints both statistically significant and non-significant edges, "sig" prints only significant edges, and "nonsig" prints only the non-significant edges.

...

Ignored.

Value

x (invisibly).

See also

Examples

model <- tna(group_regulation)
# Small number of iterations for CRAN
boot <- bootstrap(model, iter = 10)
print(boot)
#> Non-significant Edges
#> 
#>          from         to       weight    p_value     ci_lower    ci_upper
#> 2       adapt   cohesion 0.0029498525 0.45454545 0.0017550137 0.004926479
#> 3       adapt  consensus 0.0047400853 0.27272727 0.0034331332 0.006213130
#> 4       adapt coregulate 0.0162436548 0.18181818 0.0128675721 0.020885417
#> 5       adapt    discuss 0.0713743356 0.09090909 0.0687983040 0.078822516
#> 6       adapt    emotion 0.0024673951 0.36363636 0.0021505918 0.004391937
#> 7       adapt    monitor 0.0111653873 0.36363636 0.0095491365 0.015587040
#> 8       adapt       plan 0.0009745006 0.54545455 0.0008056013 0.001942620
#> 9       adapt  synthesis 0.2346625767 0.09090909 0.2073847227 0.262155785
#> 10   cohesion      adapt 0.2730844794 0.09090909 0.2442341189 0.329942272
#> 11   cohesion   cohesion 0.0271386431 0.09090909 0.0238703206 0.032505892
#> 12   cohesion  consensus 0.0148522673 0.09090909 0.0132377440 0.017681604
#> 13   cohesion coregulate 0.0360406091 0.09090909 0.0299888740 0.044009905
#> 14   cohesion    discuss 0.0475828904 0.09090909 0.0451923519 0.052796237
#> 15   cohesion    emotion 0.3253436729 0.09090909 0.3119549629 0.333068389
#> 16   cohesion    monitor 0.0558269365 0.09090909 0.0487216917 0.066299245
#> 17   cohesion       plan 0.0251745980 0.09090909 0.0218618428 0.027340841
#> 18   cohesion  synthesis 0.0337423313 0.36363636 0.0187106093 0.042546325
#> 19  consensus      adapt 0.4774066798 0.09090909 0.4426734491 0.517424242
#> 20  consensus   cohesion 0.4979351032 0.09090909 0.4743573025 0.515906565
#> 21  consensus  consensus 0.0820034761 0.09090909 0.0765718045 0.083304923
#> 22  consensus coregulate 0.1345177665 0.09090909 0.1225713401 0.145915197
#> 23  consensus    discuss 0.3211845103 0.09090909 0.3050614045 0.327916211
#> 24  consensus    emotion 0.3204088826 0.09090909 0.3125679905 0.327573621
#> 25  consensus    monitor 0.1591067690 0.09090909 0.1339142954 0.168560877
#> 26  consensus       plan 0.2904011694 0.09090909 0.2818148380 0.300710469
#> 27  consensus  synthesis 0.4662576687 0.09090909 0.4237739048 0.518855125
#> 28 coregulate      adapt 0.0216110020 0.36363636 0.0162596924 0.037723244
#> 29 coregulate   cohesion 0.1191740413 0.09090909 0.1028003374 0.135239078
#> 30 coregulate  consensus 0.1877073787 0.09090909 0.1831810864 0.190807835
#> 31 coregulate coregulate 0.0233502538 0.18181818 0.0216455919 0.029151342
#> 32 coregulate    discuss 0.0842824601 0.09090909 0.0765428605 0.094615301
#> 33 coregulate    emotion 0.0341910469 0.09090909 0.0278640102 0.038864016
#> 34 coregulate    monitor 0.0579204466 0.09090909 0.0519238155 0.064359617
#> 35 coregulate       plan 0.0172161767 0.09090909 0.0148844262 0.018912174
#> 36 coregulate  synthesis 0.0444785276 0.36363636 0.0305046744 0.064914702
#> 37    discuss      adapt 0.0589390963 0.36363636 0.0395203317 0.082817002
#> 38    discuss   cohesion 0.0595870206 0.09090909 0.0503392441 0.064522432
#> 39    discuss  consensus 0.1880233844 0.09090909 0.1867290648 0.194531563
#> 40    discuss coregulate 0.2736040609 0.09090909 0.2590429766 0.279047882
#> 41    discuss    discuss 0.1948873703 0.09090909 0.1905085569 0.204169675
#> 42    discuss    emotion 0.1018681706 0.09090909 0.0964026295 0.110459814
#> 43    discuss    monitor 0.3754361479 0.09090909 0.3522781744 0.388264025
#> 44    discuss       plan 0.0678902063 0.09090909 0.0635431530 0.071107104
#> 45    discuss  synthesis 0.0628834356 0.27272727 0.0467828475 0.084203829
#> 46    emotion      adapt 0.1198428291 0.09090909 0.1012218514 0.135770029
#> 47    emotion   cohesion 0.1156342183 0.09090909 0.1039210764 0.130812612
#> 48    emotion  consensus 0.0726813083 0.09090909 0.0676362565 0.077129429
#> 49    emotion coregulate 0.1720812183 0.09090909 0.1522867560 0.184646811
#> 50    emotion    discuss 0.1057960010 0.09090909 0.1015622032 0.115238289
#> 51    emotion    emotion 0.0768417342 0.09090909 0.0651441417 0.082297072
#> 52    emotion    monitor 0.0907187718 0.09090909 0.0824075321 0.103472558
#> 53    emotion       plan 0.1468247523 0.09090909 0.1421744803 0.150436324
#> 54    emotion  synthesis 0.0705521472 0.09090909 0.0634985157 0.076559587
#> 55    monitor      adapt 0.0333988212 0.27272727 0.0283873644 0.053488136
#> 56    monitor   cohesion 0.0330383481 0.09090909 0.0258735732 0.037088877
#> 57    monitor  consensus 0.0466108390 0.09090909 0.0442416537 0.050004431
#> 58    monitor coregulate 0.0862944162 0.09090909 0.0777997654 0.092707590
#> 59    monitor    discuss 0.0222728423 0.18181818 0.0165662331 0.023024561
#> 60    monitor    emotion 0.0363059570 0.09090909 0.0311377806 0.043999895
#> 61    monitor    monitor 0.0181437544 0.27272727 0.0169098659 0.024263159
#> 62    monitor       plan 0.0755237941 0.09090909 0.0686769820 0.080839662
#> 63    monitor  synthesis 0.0122699387 0.54545455 0.0038742284 0.015247496
#> 64       plan      adapt 0.0157170923 0.45454545 0.0086708309 0.027746212
#> 65       plan   cohesion 0.1410029499 0.09090909 0.1313348400 0.146565453
#> 66       plan  consensus 0.3957971243 0.09090909 0.3866315720 0.400136260
#> 67       plan coregulate 0.2390862944 0.09090909 0.2244951277 0.262233381
#> 68       plan    discuss 0.0116426221 0.09090909 0.0106925581 0.014300646
#> 69       plan    emotion 0.0997532605 0.09090909 0.0889018332 0.104882725
#> 70       plan    monitor 0.2156315422 0.09090909 0.1985826827 0.230606833
#> 71       plan       plan 0.3742082183 0.09090909 0.3638816573 0.388413163
#> 72       plan  synthesis 0.0751533742 0.09090909 0.0728398157 0.087949216
#> 74  synthesis   cohesion 0.0035398230 0.36363636 0.0020217594 0.005659976
#> 75  synthesis  consensus 0.0075841365 0.27272727 0.0056413711 0.010242796
#> 76  synthesis coregulate 0.0187817259 0.36363636 0.0128274116 0.024732357
#> 77  synthesis    discuss 0.1409769679 0.09090909 0.1334383851 0.146272784
#> 78  synthesis    emotion 0.0028198802 0.54545455 0.0017579660 0.005410704
#> 79  synthesis    monitor 0.0160502442 0.63636364 0.0109092251 0.024338140
#> 80  synthesis       plan 0.0017865844 0.63636364 0.0011377194 0.002416022