Print Bootstrap Results

Print Bootstrap Results

Usage

# S3 method for class 'tna_bootstrap'
print(x, digits = getOption("digits"), type = "both", ...)

Arguments

x

A tna_bootstrap object.

digits

An integer giving the minimal number of significant digits to print.

type

A character vector giving the type of edges to print. The default option "both" prints both statistically significant and non-significant edges, "sig" prints only significant edges, and "nonsig" prints only the non-significant edges.

...

Ignored.

Value

x (invisibly).

See also

Examples

model <- tna(group_regulation)
# Small number of iterations for CRAN
boot <- bootstrap(model, iter = 10)
print(boot)
#> Non-significant Edges
#> 
#>          from         to       weight    p_value     cr_lower    cr_upper
#> 2    cohesion      adapt 0.0029498525 0.63636364 0.0022123894 0.003687316
#> 3   consensus      adapt 0.0047400853 0.27272727 0.0035550640 0.005925107
#> 4  coregulate      adapt 0.0162436548 0.18181818 0.0121827411 0.020304569
#> 5     discuss      adapt 0.0713743356 0.09090909 0.0535307517 0.089217920
#> 6     emotion      adapt 0.0024673951 0.81818182 0.0018505464 0.003084244
#> 7     monitor      adapt 0.0111653873 0.36363636 0.0083740405 0.013956734
#> 8        plan      adapt 0.0009745006 0.63636364 0.0007308754 0.001218126
#> 9   synthesis      adapt 0.2346625767 0.09090909 0.1759969325 0.293328221
#> 10      adapt   cohesion 0.2730844794 0.09090909 0.2048133595 0.341355599
#> 11   cohesion   cohesion 0.0271386431 0.18181818 0.0203539823 0.033923304
#> 12  consensus   cohesion 0.0148522673 0.09090909 0.0111392005 0.018565334
#> 13 coregulate   cohesion 0.0360406091 0.18181818 0.0270304569 0.045050761
#> 14    discuss   cohesion 0.0475828904 0.09090909 0.0356871678 0.059478613
#> 15    emotion   cohesion 0.3253436729 0.09090909 0.2440077547 0.406679591
#> 16    monitor   cohesion 0.0558269365 0.09090909 0.0418702024 0.069783671
#> 17       plan   cohesion 0.0251745980 0.09090909 0.0188809485 0.031468248
#> 18  synthesis   cohesion 0.0337423313 0.27272727 0.0253067485 0.042177914
#> 19      adapt  consensus 0.4774066798 0.09090909 0.3580550098 0.596758350
#> 20   cohesion  consensus 0.4979351032 0.09090909 0.3734513274 0.622418879
#> 21  consensus  consensus 0.0820034761 0.09090909 0.0615026070 0.102504345
#> 22 coregulate  consensus 0.1345177665 0.09090909 0.1008883249 0.168147208
#> 23    discuss  consensus 0.3211845103 0.09090909 0.2408883827 0.401480638
#> 24    emotion  consensus 0.3204088826 0.09090909 0.2403066620 0.400511103
#> 25    monitor  consensus 0.1591067690 0.09090909 0.1193300768 0.198883461
#> 26       plan  consensus 0.2904011694 0.09090909 0.2178008771 0.363001462
#> 27  synthesis  consensus 0.4662576687 0.09090909 0.3496932515 0.582822086
#> 28      adapt coregulate 0.0216110020 0.36363636 0.0162082515 0.027013752
#> 29   cohesion coregulate 0.1191740413 0.09090909 0.0893805310 0.148967552
#> 30  consensus coregulate 0.1877073787 0.09090909 0.1407805340 0.234634223
#> 31 coregulate coregulate 0.0233502538 0.18181818 0.0175126904 0.029187817
#> 32    discuss coregulate 0.0842824601 0.09090909 0.0632118451 0.105353075
#> 33    emotion coregulate 0.0341910469 0.09090909 0.0256432852 0.042738809
#> 34    monitor coregulate 0.0579204466 0.09090909 0.0434403350 0.072400558
#> 35       plan coregulate 0.0172161767 0.09090909 0.0129121325 0.021520221
#> 36  synthesis coregulate 0.0444785276 0.36363636 0.0333588957 0.055598160
#> 37      adapt    discuss 0.0589390963 0.27272727 0.0442043222 0.073673870
#> 38   cohesion    discuss 0.0595870206 0.09090909 0.0446902655 0.074483776
#> 39  consensus    discuss 0.1880233844 0.09090909 0.1410175383 0.235029231
#> 40 coregulate    discuss 0.2736040609 0.09090909 0.2052030457 0.342005076
#> 41    discuss    discuss 0.1948873703 0.09090909 0.1461655277 0.243609213
#> 42    emotion    discuss 0.1018681706 0.09090909 0.0764011280 0.127335213
#> 43    monitor    discuss 0.3754361479 0.09090909 0.2815771110 0.469295185
#> 44       plan    discuss 0.0678902063 0.09090909 0.0509176547 0.084862758
#> 45  synthesis    discuss 0.0628834356 0.09090909 0.0471625767 0.078604294
#> 46      adapt    emotion 0.1198428291 0.09090909 0.0898821218 0.149803536
#> 47   cohesion    emotion 0.1156342183 0.09090909 0.0867256637 0.144542773
#> 48  consensus    emotion 0.0726813083 0.09090909 0.0545109812 0.090851635
#> 49 coregulate    emotion 0.1720812183 0.09090909 0.1290609137 0.215101523
#> 50    discuss    emotion 0.1057960010 0.09090909 0.0793470008 0.132245001
#> 51    emotion    emotion 0.0768417342 0.09090909 0.0576313007 0.096052168
#> 52    monitor    emotion 0.0907187718 0.09090909 0.0680390789 0.113398465
#> 53       plan    emotion 0.1468247523 0.09090909 0.1101185642 0.183530940
#> 54  synthesis    emotion 0.0705521472 0.18181818 0.0529141104 0.088190184
#> 55      adapt    monitor 0.0333988212 0.36363636 0.0250491159 0.041748527
#> 56   cohesion    monitor 0.0330383481 0.18181818 0.0247787611 0.041297935
#> 57  consensus    monitor 0.0466108390 0.09090909 0.0349581292 0.058263549
#> 58 coregulate    monitor 0.0862944162 0.09090909 0.0647208122 0.107868020
#> 59    discuss    monitor 0.0222728423 0.09090909 0.0167046317 0.027841053
#> 60    emotion    monitor 0.0363059570 0.09090909 0.0272294677 0.045382446
#> 61    monitor    monitor 0.0181437544 0.27272727 0.0136078158 0.022679693
#> 62       plan    monitor 0.0755237941 0.09090909 0.0566428455 0.094404743
#> 63  synthesis    monitor 0.0122699387 0.45454545 0.0092024540 0.015337423
#> 64      adapt       plan 0.0157170923 0.72727273 0.0117878193 0.019646365
#> 65   cohesion       plan 0.1410029499 0.09090909 0.1057522124 0.176253687
#> 66  consensus       plan 0.3957971243 0.09090909 0.2968478433 0.494746405
#> 67 coregulate       plan 0.2390862944 0.09090909 0.1793147208 0.298857868
#> 68    discuss       plan 0.0116426221 0.09090909 0.0087319666 0.014553278
#> 69    emotion       plan 0.0997532605 0.09090909 0.0748149454 0.124691576
#> 70    monitor       plan 0.2156315422 0.09090909 0.1617236567 0.269539428
#> 71       plan       plan 0.3742082183 0.09090909 0.2806561637 0.467760273
#> 72  synthesis       plan 0.0751533742 0.18181818 0.0563650307 0.093941718
#> 74   cohesion  synthesis 0.0035398230 0.54545455 0.0026548673 0.004424779
#> 75  consensus  synthesis 0.0075841365 0.18181818 0.0056881024 0.009480171
#> 76 coregulate  synthesis 0.0187817259 0.18181818 0.0140862944 0.023477157
#> 77    discuss  synthesis 0.1409769679 0.09090909 0.1057327259 0.176221210
#> 78    emotion  synthesis 0.0028198802 0.27272727 0.0021149101 0.003524850
#> 79    monitor  synthesis 0.0160502442 0.27272727 0.0120376832 0.020062805
#> 80       plan  synthesis 0.0017865844 0.72727273 0.0013399383 0.002233230
#>        ci_lower    ci_upper
#> 2  0.0011865566 0.005102214
#> 3  0.0032314850 0.005998093
#> 4  0.0118679001 0.018580886
#> 5  0.0655236546 0.079896383
#> 6  0.0007602321 0.003921587
#> 7  0.0087608964 0.017460336
#> 8  0.0005130727 0.001549549
#> 9  0.2098110862 0.266986279
#> 10 0.2480593817 0.300314932
#> 11 0.0191343723 0.029844715
#> 12 0.0129000114 0.016350573
#> 13 0.0295879221 0.044419379
#> 14 0.0388467021 0.054629082
#> 15 0.3145005096 0.346053719
#> 16 0.0450692785 0.061204865
#> 17 0.0244376090 0.029278209
#> 18 0.0181203931 0.044219961
#> 19 0.4468969444 0.505740741
#> 20 0.4764470329 0.517993357
#> 21 0.0780935635 0.089793987
#> 22 0.1276859765 0.148017665
#> 23 0.3143923037 0.334977713
#> 24 0.3017117351 0.329977485
#> 25 0.1511064022 0.169141404
#> 26 0.2844877632 0.298617275
#> 27 0.4447470194 0.489709596
#> 28 0.0163100764 0.035190190
#> 29 0.1065144587 0.138546849
#> 30 0.1801515367 0.194002511
#> 31 0.0187313491 0.028666457
#> 32 0.0745633526 0.091072100
#> 33 0.0306254489 0.039863692
#> 34 0.0535140780 0.069638367
#> 35 0.0151664659 0.019302813
#> 36 0.0324325505 0.068410578
#> 37 0.0413241306 0.073890518
#> 38 0.0528869074 0.064763291
#> 39 0.1819387417 0.199396953
#> 40 0.2595958743 0.286207879
#> 41 0.1837964942 0.201759478
#> 42 0.0923537928 0.111266469
#> 43 0.3654915446 0.389890975
#> 44 0.0629202899 0.070613349
#> 45 0.0545542636 0.077087697
#> 46 0.1011280531 0.136196179
#> 47 0.1016163251 0.131444058
#> 48 0.0691429848 0.077267015
#> 49 0.1538236366 0.181417556
#> 50 0.1036541892 0.110722861
#> 51 0.0675996082 0.083325693
#> 52 0.0762379019 0.102541295
#> 53 0.1413629186 0.152299448
#> 54 0.0543918919 0.086135431
#> 55 0.0230067568 0.045822262
#> 56 0.0245950424 0.040246992
#> 57 0.0406017549 0.047638358
#> 58 0.0767235884 0.090581577
#> 59 0.0190994711 0.025540587
#> 60 0.0315498630 0.041976021
#> 61 0.0133653480 0.024778827
#> 62 0.0709359990 0.078640604
#> 63 0.0080166805 0.021209476
#> 64 0.0037925015 0.019807001
#> 65 0.1226636292 0.151856893
#> 66 0.3950605671 0.406436912
#> 67 0.2283599381 0.257761639
#> 68 0.0094547100 0.014177996
#> 69 0.0927540570 0.106569302
#> 70 0.2030516509 0.231007057
#> 71 0.3671373824 0.383754023
#> 72 0.0545803917 0.084021279
#> 74 0.0017842413 0.005531693
#> 75 0.0066734976 0.009550029
#> 76 0.0133920612 0.023033351
#> 77 0.1336548198 0.150043923
#> 78 0.0015069052 0.003405250
#> 79 0.0099984210 0.020156352
#> 80 0.0011151712 0.002469873