Print Bootstrap Results

Print Bootstrap Results

Usage

# S3 method for class 'tna_bootstrap'
print(x, digits = getOption("digits"), type = "both", ...)

Arguments

x

A tna_bootstrap object.

digits

An integer giving the minimal number of significant digits to print.

type

A character vector giving the type of edges to print. The default option "both" prints both statistically significant and non-significant edges, "sig" prints only significant edges, and "nonsig" prints only the non-significant edges.

...

Ignored.

Value

x (invisibly).

Examples

model <- tna(group_regulation)
# Small number of iterations for CRAN
boot <- bootstrap(model, iter = 10)
print(boot)
#> Significant Edges
#> 
#>          from         to     weight p_value   ci_lower   ci_upper
#> 5       adapt    discuss 0.07137434       0 0.06348657 0.08099424
#> 9       adapt  synthesis 0.23466258       0 0.21272779 0.23699189
#> 10   cohesion      adapt 0.27308448       0 0.26305384 0.30787697
#> 11   cohesion   cohesion 0.02713864       0 0.02330384 0.03177955
#> 13   cohesion coregulate 0.03604061       0 0.02987657 0.04196894
#> 14   cohesion    discuss 0.04758289       0 0.04230633 0.05101607
#> 15   cohesion    emotion 0.32534367       0 0.31056561 0.33851359
#> 16   cohesion    monitor 0.05582694       0 0.04776232 0.06208833
#> 17   cohesion       plan 0.02517460       0 0.02128413 0.03057810
#> 19  consensus      adapt 0.47740668       0 0.45386127 0.48493183
#> 20  consensus   cohesion 0.49793510       0 0.48458249 0.50812328
#> 21  consensus  consensus 0.08200348       0 0.07800969 0.08618305
#> 22  consensus coregulate 0.13451777       0 0.12929013 0.14862154
#> 23  consensus    discuss 0.32118451       0 0.30804456 0.32536524
#> 24  consensus    emotion 0.32040888       0 0.30450104 0.33694538
#> 25  consensus    monitor 0.15910677       0 0.14920041 0.17468452
#> 26  consensus       plan 0.29040117       0 0.28456932 0.30239099
#> 27  consensus  synthesis 0.46625767       0 0.44943706 0.49272703
#> 29 coregulate   cohesion 0.11917404       0 0.10535095 0.12872533
#> 30 coregulate  consensus 0.18770738       0 0.17960516 0.19951225
#> 32 coregulate    discuss 0.08428246       0 0.07729910 0.09040363
#> 33 coregulate    emotion 0.03419105       0 0.03050117 0.03768501
#> 34 coregulate    monitor 0.05792045       0 0.04696763 0.06337494
#> 35 coregulate       plan 0.01721618       0 0.01532915 0.01929869
#> 38    discuss   cohesion 0.05958702       0 0.04926160 0.06504461
#> 39    discuss  consensus 0.18802338       0 0.17868232 0.19806794
#> 40    discuss coregulate 0.27360406       0 0.26904263 0.29244461
#> 41    discuss    discuss 0.19488737       0 0.18202403 0.19559620
#> 42    discuss    emotion 0.10186817       0 0.09235798 0.11181148
#> 43    discuss    monitor 0.37543615       0 0.36581044 0.39827111
#> 44    discuss       plan 0.06789021       0 0.06073033 0.07115174
#> 47    emotion   cohesion 0.11563422       0 0.10105875 0.12442661
#> 48    emotion  consensus 0.07268131       0 0.06595059 0.07555192
#> 49    emotion coregulate 0.17208122       0 0.15258146 0.18469107
#> 50    emotion    discuss 0.10579600       0 0.09994651 0.11380820
#> 51    emotion    emotion 0.07684173       0 0.06976067 0.08241250
#> 52    emotion    monitor 0.09071877       0 0.08490015 0.09816890
#> 53    emotion       plan 0.14682475       0 0.14062275 0.15338833
#> 54    emotion  synthesis 0.07055215       0 0.06007466 0.08295926
#> 56    monitor   cohesion 0.03303835       0 0.03011232 0.03828852
#> 57    monitor  consensus 0.04661084       0 0.04382291 0.05042065
#> 58    monitor coregulate 0.08629442       0 0.07927459 0.09466942
#> 60    monitor    emotion 0.03630596       0 0.03362938 0.04010115
#> 62    monitor       plan 0.07552379       0 0.07107939 0.07922775
#> 65       plan   cohesion 0.14100295       0 0.13168132 0.14950035
#> 66       plan  consensus 0.39579712       0 0.38210235 0.40138769
#> 67       plan coregulate 0.23908629       0 0.22246828 0.25328213
#> 69       plan    emotion 0.09975326       0 0.09060304 0.10546900
#> 70       plan    monitor 0.21563154       0 0.19869750 0.22991653
#> 71       plan       plan 0.37420822       0 0.36775016 0.38409616
#> 72       plan  synthesis 0.07515337       0 0.06096315 0.08950124
#> 76  synthesis coregulate 0.01878173       0 0.01669537 0.02125180
#> 77  synthesis    discuss 0.14097697       0 0.13674126 0.15157504
#> 
#> Non-significant Edges
#> 
#>          from         to       weight p_value     ci_lower    ci_upper
#> 2       adapt   cohesion 0.0029498525     0.6 0.0011700722 0.005528411
#> 3       adapt  consensus 0.0047400853     0.3 0.0028521662 0.005868705
#> 4       adapt coregulate 0.0162436548     0.2 0.0103461444 0.021540675
#> 6       adapt    emotion 0.0024673951     0.5 0.0011637313 0.004472447
#> 7       adapt    monitor 0.0111653873     0.3 0.0081916230 0.014171023
#> 8       adapt       plan 0.0009745006     0.5 0.0005194638 0.001591307
#> 12   cohesion  consensus 0.0148522673     0.1 0.0124890674 0.018225139
#> 18   cohesion  synthesis 0.0337423313     0.4 0.0212611070 0.047220972
#> 28 coregulate      adapt 0.0216110020     0.4 0.0091991976 0.033621274
#> 31 coregulate coregulate 0.0233502538     0.2 0.0171628787 0.028761038
#> 36 coregulate  synthesis 0.0444785276     0.2 0.0417210641 0.060518041
#> 37    discuss      adapt 0.0589390963     0.2 0.0440868579 0.073832177
#> 45    discuss  synthesis 0.0628834356     0.2 0.0435760679 0.078226828
#> 46    emotion      adapt 0.1198428291     0.1 0.0937062498 0.146211234
#> 55    monitor      adapt 0.0333988212     0.4 0.0262025503 0.052730413
#> 59    monitor    discuss 0.0222728423     0.1 0.0185202089 0.027612504
#> 61    monitor    monitor 0.0181437544     0.1 0.0139352029 0.021040082
#> 63    monitor  synthesis 0.0122699387     0.5 0.0063888520 0.019261377
#> 64       plan      adapt 0.0157170923     0.4 0.0053042228 0.024514390
#> 68       plan    discuss 0.0116426221     0.1 0.0099493880 0.015457585
#> 74  synthesis   cohesion 0.0035398230     0.4 0.0023272497 0.004929258
#> 75  synthesis  consensus 0.0075841365     0.1 0.0054522797 0.008799488
#> 78  synthesis    emotion 0.0028198802     0.5 0.0012470765 0.004039703
#> 79  synthesis    monitor 0.0160502442     0.4 0.0092235087 0.019968951
#> 80  synthesis       plan 0.0017865844     0.3 0.0010049210 0.002443799