Print Bootstrap Results

Print Bootstrap Results

Usage

# S3 method for class 'tna_bootstrap'
print(x, digits = getOption("digits"), type = "both", ...)

Arguments

x

A tna_bootstrap object.

digits

An integer giving the minimal number of significant digits to print.

type

A character vector giving the type of edges to print. The default option "both" prints both statistically significant and non-significant edges, "sig" prints only significant edges, and "nonsig" prints only the non-significant edges.

...

Ignored.

Value

x (invisibly).

See also

Examples

model <- tna(group_regulation)
# Small number of iterations for CRAN
boot <- bootstrap(model, iter = 10)
print(boot)
#> Significant Edges
#> 
#>          from         to     weight p_value   ci_lower   ci_upper
#> 5       adapt    discuss 0.07137434       0 0.06879830 0.07882252
#> 9       adapt  synthesis 0.23466258       0 0.20738472 0.26215579
#> 10   cohesion      adapt 0.27308448       0 0.24423412 0.32994227
#> 11   cohesion   cohesion 0.02713864       0 0.02387032 0.03250589
#> 12   cohesion  consensus 0.01485227       0 0.01323774 0.01768160
#> 13   cohesion coregulate 0.03604061       0 0.02998887 0.04400990
#> 14   cohesion    discuss 0.04758289       0 0.04519235 0.05279624
#> 15   cohesion    emotion 0.32534367       0 0.31195496 0.33306839
#> 16   cohesion    monitor 0.05582694       0 0.04872169 0.06629924
#> 17   cohesion       plan 0.02517460       0 0.02186184 0.02734084
#> 19  consensus      adapt 0.47740668       0 0.44267345 0.51742424
#> 20  consensus   cohesion 0.49793510       0 0.47435730 0.51590656
#> 21  consensus  consensus 0.08200348       0 0.07657180 0.08330492
#> 22  consensus coregulate 0.13451777       0 0.12257134 0.14591520
#> 23  consensus    discuss 0.32118451       0 0.30506140 0.32791621
#> 24  consensus    emotion 0.32040888       0 0.31256799 0.32757362
#> 25  consensus    monitor 0.15910677       0 0.13391430 0.16856088
#> 26  consensus       plan 0.29040117       0 0.28181484 0.30071047
#> 27  consensus  synthesis 0.46625767       0 0.42377390 0.51885512
#> 29 coregulate   cohesion 0.11917404       0 0.10280034 0.13523908
#> 30 coregulate  consensus 0.18770738       0 0.18318109 0.19080783
#> 32 coregulate    discuss 0.08428246       0 0.07654286 0.09461530
#> 33 coregulate    emotion 0.03419105       0 0.02786401 0.03886402
#> 34 coregulate    monitor 0.05792045       0 0.05192382 0.06435962
#> 35 coregulate       plan 0.01721618       0 0.01488443 0.01891217
#> 38    discuss   cohesion 0.05958702       0 0.05033924 0.06452243
#> 39    discuss  consensus 0.18802338       0 0.18672906 0.19453156
#> 40    discuss coregulate 0.27360406       0 0.25904298 0.27904788
#> 41    discuss    discuss 0.19488737       0 0.19050856 0.20416967
#> 42    discuss    emotion 0.10186817       0 0.09640263 0.11045981
#> 43    discuss    monitor 0.37543615       0 0.35227817 0.38826403
#> 44    discuss       plan 0.06789021       0 0.06354315 0.07110710
#> 46    emotion      adapt 0.11984283       0 0.10122185 0.13577003
#> 47    emotion   cohesion 0.11563422       0 0.10392108 0.13081261
#> 48    emotion  consensus 0.07268131       0 0.06763626 0.07712943
#> 49    emotion coregulate 0.17208122       0 0.15228676 0.18464681
#> 50    emotion    discuss 0.10579600       0 0.10156220 0.11523829
#> 51    emotion    emotion 0.07684173       0 0.06514414 0.08229707
#> 52    emotion    monitor 0.09071877       0 0.08240753 0.10347256
#> 53    emotion       plan 0.14682475       0 0.14217448 0.15043632
#> 54    emotion  synthesis 0.07055215       0 0.06349852 0.07655959
#> 56    monitor   cohesion 0.03303835       0 0.02587357 0.03708888
#> 57    monitor  consensus 0.04661084       0 0.04424165 0.05000443
#> 58    monitor coregulate 0.08629442       0 0.07779977 0.09270759
#> 60    monitor    emotion 0.03630596       0 0.03113778 0.04399990
#> 62    monitor       plan 0.07552379       0 0.06867698 0.08083966
#> 65       plan   cohesion 0.14100295       0 0.13133484 0.14656545
#> 66       plan  consensus 0.39579712       0 0.38663157 0.40013626
#> 67       plan coregulate 0.23908629       0 0.22449513 0.26223338
#> 68       plan    discuss 0.01164262       0 0.01069256 0.01430065
#> 69       plan    emotion 0.09975326       0 0.08890183 0.10488273
#> 70       plan    monitor 0.21563154       0 0.19858268 0.23060683
#> 71       plan       plan 0.37420822       0 0.36388166 0.38841316
#> 72       plan  synthesis 0.07515337       0 0.07283982 0.08794922
#> 77  synthesis    discuss 0.14097697       0 0.13343839 0.14627278
#> 
#> Non-significant Edges
#> 
#>          from         to       weight p_value     ci_lower    ci_upper
#> 2       adapt   cohesion 0.0029498525     0.4 0.0017550137 0.004926479
#> 3       adapt  consensus 0.0047400853     0.2 0.0034331332 0.006213130
#> 4       adapt coregulate 0.0162436548     0.1 0.0128675721 0.020885417
#> 6       adapt    emotion 0.0024673951     0.3 0.0021505918 0.004391937
#> 7       adapt    monitor 0.0111653873     0.3 0.0095491365 0.015587040
#> 8       adapt       plan 0.0009745006     0.5 0.0008056013 0.001942620
#> 18   cohesion  synthesis 0.0337423313     0.3 0.0187106093 0.042546325
#> 28 coregulate      adapt 0.0216110020     0.3 0.0162596924 0.037723244
#> 31 coregulate coregulate 0.0233502538     0.1 0.0216455919 0.029151342
#> 36 coregulate  synthesis 0.0444785276     0.3 0.0305046744 0.064914702
#> 37    discuss      adapt 0.0589390963     0.3 0.0395203317 0.082817002
#> 45    discuss  synthesis 0.0628834356     0.2 0.0467828475 0.084203829
#> 55    monitor      adapt 0.0333988212     0.2 0.0283873644 0.053488136
#> 59    monitor    discuss 0.0222728423     0.1 0.0165662331 0.023024561
#> 61    monitor    monitor 0.0181437544     0.2 0.0169098659 0.024263159
#> 63    monitor  synthesis 0.0122699387     0.5 0.0038742284 0.015247496
#> 64       plan      adapt 0.0157170923     0.4 0.0086708309 0.027746212
#> 74  synthesis   cohesion 0.0035398230     0.3 0.0020217594 0.005659976
#> 75  synthesis  consensus 0.0075841365     0.2 0.0056413711 0.010242796
#> 76  synthesis coregulate 0.0187817259     0.3 0.0128274116 0.024732357
#> 78  synthesis    emotion 0.0028198802     0.5 0.0017579660 0.005410704
#> 79  synthesis    monitor 0.0160502442     0.6 0.0109092251 0.024338140
#> 80  synthesis       plan 0.0017865844     0.6 0.0011377194 0.002416022