Print Bootstrap Results

Print Bootstrap Results

Usage

# S3 method for class 'tna_bootstrap'
print(x, digits = getOption("digits"), type = "both", ...)

Arguments

x

A tna_bootstrap object.

digits

An integer giving the minimal number of significant digits to print.

type

A character vector giving the type of edges to print. The default option "both" prints both statistically significant and non-significant edges, "sig" prints only significant edges, and "nonsig" prints only the non-significant edges.

...

Ignored.

Value

x (invisibly).

See also

Examples

model <- tna(group_regulation)
# Small number of iterations for CRAN
boot <- bootstrap(model, iter = 10)
print(boot)
#> Non-significant Edges
#> 
#>          from         to       weight    p_value     cr_lower    cr_upper
#> 2    cohesion      adapt 0.0029498525 0.90909091 0.0022123894 0.003687316
#> 3   consensus      adapt 0.0047400853 0.27272727 0.0035550640 0.005925107
#> 4  coregulate      adapt 0.0162436548 0.27272727 0.0121827411 0.020304569
#> 5     discuss      adapt 0.0713743356 0.09090909 0.0535307517 0.089217920
#> 6     emotion      adapt 0.0024673951 0.36363636 0.0018505464 0.003084244
#> 7     monitor      adapt 0.0111653873 0.36363636 0.0083740405 0.013956734
#> 8        plan      adapt 0.0009745006 0.63636364 0.0007308754 0.001218126
#> 9   synthesis      adapt 0.2346625767 0.09090909 0.1759969325 0.293328221
#> 10      adapt   cohesion 0.2730844794 0.09090909 0.2048133595 0.341355599
#> 11   cohesion   cohesion 0.0271386431 0.27272727 0.0203539823 0.033923304
#> 12  consensus   cohesion 0.0148522673 0.09090909 0.0111392005 0.018565334
#> 13 coregulate   cohesion 0.0360406091 0.18181818 0.0270304569 0.045050761
#> 14    discuss   cohesion 0.0475828904 0.09090909 0.0356871678 0.059478613
#> 15    emotion   cohesion 0.3253436729 0.09090909 0.2440077547 0.406679591
#> 16    monitor   cohesion 0.0558269365 0.09090909 0.0418702024 0.069783671
#> 17       plan   cohesion 0.0251745980 0.09090909 0.0188809485 0.031468248
#> 18  synthesis   cohesion 0.0337423313 0.36363636 0.0253067485 0.042177914
#> 19      adapt  consensus 0.4774066798 0.09090909 0.3580550098 0.596758350
#> 20   cohesion  consensus 0.4979351032 0.09090909 0.3734513274 0.622418879
#> 21  consensus  consensus 0.0820034761 0.09090909 0.0615026070 0.102504345
#> 22 coregulate  consensus 0.1345177665 0.09090909 0.1008883249 0.168147208
#> 23    discuss  consensus 0.3211845103 0.09090909 0.2408883827 0.401480638
#> 24    emotion  consensus 0.3204088826 0.09090909 0.2403066620 0.400511103
#> 25    monitor  consensus 0.1591067690 0.09090909 0.1193300768 0.198883461
#> 26       plan  consensus 0.2904011694 0.09090909 0.2178008771 0.363001462
#> 27  synthesis  consensus 0.4662576687 0.09090909 0.3496932515 0.582822086
#> 28      adapt coregulate 0.0216110020 0.45454545 0.0162082515 0.027013752
#> 29   cohesion coregulate 0.1191740413 0.09090909 0.0893805310 0.148967552
#> 30  consensus coregulate 0.1877073787 0.09090909 0.1407805340 0.234634223
#> 31 coregulate coregulate 0.0233502538 0.09090909 0.0175126904 0.029187817
#> 32    discuss coregulate 0.0842824601 0.09090909 0.0632118451 0.105353075
#> 33    emotion coregulate 0.0341910469 0.09090909 0.0256432852 0.042738809
#> 34    monitor coregulate 0.0579204466 0.09090909 0.0434403350 0.072400558
#> 35       plan coregulate 0.0172161767 0.09090909 0.0129121325 0.021520221
#> 36  synthesis coregulate 0.0444785276 0.54545455 0.0333588957 0.055598160
#> 37      adapt    discuss 0.0589390963 0.18181818 0.0442043222 0.073673870
#> 38   cohesion    discuss 0.0595870206 0.09090909 0.0446902655 0.074483776
#> 39  consensus    discuss 0.1880233844 0.09090909 0.1410175383 0.235029231
#> 40 coregulate    discuss 0.2736040609 0.09090909 0.2052030457 0.342005076
#> 41    discuss    discuss 0.1948873703 0.09090909 0.1461655277 0.243609213
#> 42    emotion    discuss 0.1018681706 0.09090909 0.0764011280 0.127335213
#> 43    monitor    discuss 0.3754361479 0.09090909 0.2815771110 0.469295185
#> 44       plan    discuss 0.0678902063 0.09090909 0.0509176547 0.084862758
#> 45  synthesis    discuss 0.0628834356 0.09090909 0.0471625767 0.078604294
#> 46      adapt    emotion 0.1198428291 0.09090909 0.0898821218 0.149803536
#> 47   cohesion    emotion 0.1156342183 0.09090909 0.0867256637 0.144542773
#> 48  consensus    emotion 0.0726813083 0.09090909 0.0545109812 0.090851635
#> 49 coregulate    emotion 0.1720812183 0.09090909 0.1290609137 0.215101523
#> 50    discuss    emotion 0.1057960010 0.09090909 0.0793470008 0.132245001
#> 51    emotion    emotion 0.0768417342 0.09090909 0.0576313007 0.096052168
#> 52    monitor    emotion 0.0907187718 0.09090909 0.0680390789 0.113398465
#> 53       plan    emotion 0.1468247523 0.09090909 0.1101185642 0.183530940
#> 54  synthesis    emotion 0.0705521472 0.18181818 0.0529141104 0.088190184
#> 55      adapt    monitor 0.0333988212 0.27272727 0.0250491159 0.041748527
#> 56   cohesion    monitor 0.0330383481 0.18181818 0.0247787611 0.041297935
#> 57  consensus    monitor 0.0466108390 0.09090909 0.0349581292 0.058263549
#> 58 coregulate    monitor 0.0862944162 0.09090909 0.0647208122 0.107868020
#> 59    discuss    monitor 0.0222728423 0.09090909 0.0167046317 0.027841053
#> 60    emotion    monitor 0.0363059570 0.09090909 0.0272294677 0.045382446
#> 61    monitor    monitor 0.0181437544 0.45454545 0.0136078158 0.022679693
#> 62       plan    monitor 0.0755237941 0.09090909 0.0566428455 0.094404743
#> 63  synthesis    monitor 0.0122699387 0.63636364 0.0092024540 0.015337423
#> 64      adapt       plan 0.0157170923 0.45454545 0.0117878193 0.019646365
#> 65   cohesion       plan 0.1410029499 0.09090909 0.1057522124 0.176253687
#> 66  consensus       plan 0.3957971243 0.09090909 0.2968478433 0.494746405
#> 67 coregulate       plan 0.2390862944 0.09090909 0.1793147208 0.298857868
#> 68    discuss       plan 0.0116426221 0.09090909 0.0087319666 0.014553278
#> 69    emotion       plan 0.0997532605 0.09090909 0.0748149454 0.124691576
#> 70    monitor       plan 0.2156315422 0.09090909 0.1617236567 0.269539428
#> 71       plan       plan 0.3742082183 0.09090909 0.2806561637 0.467760273
#> 72  synthesis       plan 0.0751533742 0.27272727 0.0563650307 0.093941718
#> 74   cohesion  synthesis 0.0035398230 0.72727273 0.0026548673 0.004424779
#> 75  consensus  synthesis 0.0075841365 0.09090909 0.0056881024 0.009480171
#> 76 coregulate  synthesis 0.0187817259 0.18181818 0.0140862944 0.023477157
#> 77    discuss  synthesis 0.1409769679 0.09090909 0.1057327259 0.176221210
#> 78    emotion  synthesis 0.0028198802 0.45454545 0.0021149101 0.003524850
#> 79    monitor  synthesis 0.0160502442 0.36363636 0.0120376832 0.020062805
#> 80       plan  synthesis 0.0017865844 0.72727273 0.0013399383 0.002233230
#>        ci_lower    ci_upper
#> 2  0.0008363668 0.005214991
#> 3  0.0035161235 0.006237612
#> 4  0.0111658480 0.019288938
#> 5  0.0664832118 0.079340752
#> 6  0.0018600931 0.003533442
#> 7  0.0066286226 0.014969136
#> 8  0.0006519716 0.001657438
#> 9  0.2014365652 0.280485918
#> 10 0.2452614146 0.297182720
#> 11 0.0203692624 0.033988503
#> 12 0.0130282480 0.016259645
#> 13 0.0284385129 0.043903258
#> 14 0.0438478347 0.052992920
#> 15 0.3134539985 0.342328896
#> 16 0.0485920473 0.060765689
#> 17 0.0244570236 0.029177845
#> 18 0.0202951930 0.043573070
#> 19 0.4413783194 0.516892140
#> 20 0.4825715804 0.518057880
#> 21 0.0798096369 0.087550702
#> 22 0.1333752885 0.149224030
#> 23 0.3126392558 0.332645175
#> 24 0.3054084255 0.338101249
#> 25 0.1436770291 0.178776784
#> 26 0.2830773068 0.296434582
#> 27 0.4298669833 0.496652354
#> 28 0.0165875161 0.031339928
#> 29 0.1020521734 0.139865705
#> 30 0.1762351784 0.195735639
#> 31 0.0194332003 0.027852437
#> 32 0.0801480012 0.087676189
#> 33 0.0312844401 0.041143776
#> 34 0.0485055494 0.063941566
#> 35 0.0142541429 0.018501823
#> 36 0.0282459196 0.059851158
#> 37 0.0359609601 0.068967959
#> 38 0.0520173021 0.064839210
#> 39 0.1836165519 0.200930926
#> 40 0.2574464271 0.290925401
#> 41 0.1828913775 0.202644987
#> 42 0.0943546292 0.108242939
#> 43 0.3588714290 0.401488731
#> 44 0.0601613893 0.069596670
#> 45 0.0520982045 0.073672195
#> 46 0.0951378897 0.138606131
#> 47 0.0998250921 0.123227234
#> 48 0.0692329266 0.078071504
#> 49 0.1593658277 0.188647429
#> 50 0.1005106464 0.114614380
#> 51 0.0697490211 0.084162027
#> 52 0.0835790693 0.097188441
#> 53 0.1408746039 0.150954291
#> 54 0.0626526445 0.087598204
#> 55 0.0283674755 0.051912135
#> 56 0.0253517256 0.036452039
#> 57 0.0420193997 0.050042023
#> 58 0.0759010268 0.094628877
#> 59 0.0171561403 0.026620960
#> 60 0.0309829219 0.042361584
#> 61 0.0125721154 0.023131077
#> 62 0.0680428796 0.082729286
#> 63 0.0093897415 0.024330565
#> 64 0.0091412391 0.017703790
#> 65 0.1225308415 0.159029347
#> 66 0.3839145526 0.407969665
#> 67 0.2225554933 0.266477308
#> 68 0.0089526514 0.013774537
#> 69 0.0933793814 0.110637984
#> 70 0.1969412905 0.229389188
#> 71 0.3682930392 0.382286111
#> 72 0.0548316272 0.095602993
#> 74 0.0019354502 0.007119862
#> 75 0.0062516683 0.008956226
#> 76 0.0140478855 0.022839901
#> 77 0.1314692830 0.149028450
#> 78 0.0020820372 0.004062474
#> 79 0.0081214851 0.018616091
#> 80 0.0012946951 0.002869136